adding builder factory for maze, factory pattern for adventurer and creature

+++ src/main/java/Demon.java
@@ -1,5 +1,5 @@
 public class Demon extends Creature{
-    public Demon(String name, int health){
+    public Demon(String name, double health){
         super(name,15);
     }
